pyproject.toml
[project]
name = "andromeda"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mediapipe>=0.10.21",
    "opencv-python>=4.11.0.86",
    "pyyaml>=6.0.2",
]
